Fixed calculation of distance. Added uploading of distance and duration  to Xplocity service.

+++ app/src/main/java/com/xplocity/xplocity/RouteNewActivity.java
@@ -466,7 +466,7 @@ public class RouteNewActivity
 
     private void updateDistance() {
         if (mService.getRoute() != null) {
-            mTxtDistance.setText(Formatter.formatDistance(mService.getRoute().distance));
+            mTxtDistance.setText(Formatter.formatDistance((int) mService.getRoute().distance));
         }
     }
